requirements.txt
discord.py
aiohttp
python-dotenv
aioftp
matplotlib
aiofiles
asyncpg
